+++ NEWS.md
@@ -2,8 +2,8 @@
 
 - *Yupana*
   - Show equation with adjusted R in scatter plot graph
-  - `sig` include all the variables from summary table
-  - In plots if number of reps is 1 the `sig` and `error` is "none"
+  - `sig` include all the variables from the summary table
+  - In plots if the number of reps is 1 the `sig` and `error` is "none"
 
 # Version 0.4.2